Update and rename README.md to Experiment - 2
EXPERIMENT 2

To study and implement C++ Program Structure (Data Types)

Aim: The objective of this experiment is to study the overall structure of a C++ program and understand various data types such as int, float, char, and double. It focuses on declaring variables, initializing them, and displaying their values to understand how data types work in C++.

Procedure:

Create a new C++ program file.

Include the necessary header files and define the main() function.

Declare variables of different data types such as int, float, char, and double.

Initialize the variables with sample values.

Use the cout statement to display the values of the variables on the console.

Compile and run the program.

7)Observe the output and note the behavior of different data types.

Conclusion: Through this experiment, the user learned about various data types in C++ and how they are used to store different kinds of data. It also helped understand memory requirements and the importance of selecting the appropriate data type for variables in programs.